Finding springs and particles methods moved
